backend/requirements.txt
fastapi==0.109.2
uvicorn==0.27.1
python-multipart==0.0.9
Pillow==10.2.0
pytest==8.0.1
httpx==0.26.0
pydantic==2.6.1
pydantic-settings==2.1.0

# Development dependencies
black==24.1.1
flake8==7.0.0
flake8-docstrings==1.7.0
isort==5.13.2
mypy==1.8.0
pylint==3.0.3
pre-commit==3.6.0

# Type stubs
types-Pillow 